refactor(metrics): extract Error into metric_collection/error.rs
